// resources/js/pages/school/data.ts
export type Role = 'admin' | 'teacher' | 'student' | 'parent';
export type Screen =
    | 'dashboard'
    | 'students'
    | 'teachers'
    | 'classes'
    | 'attendance'
    | 'grades'
    | 'homework'
    | 'fees'
    | 'addStudent';
export type Lang = 'kh' | 'en';